Refactoring: deserializer->builder
