+++ src/main/java/io/stargate/sgv2/jsonapi/api/model/command/Sortable.java
@@ -3,7 +3,6 @@ package io.stargate.sgv2.jsonapi.api.model.command;
 import io.stargate.sgv2.jsonapi.api.model.command.clause.filter.SortSpec;
 import io.stargate.sgv2.jsonapi.api.model.command.clause.sort.SortClause;
 import io.stargate.sgv2.jsonapi.service.cqldriver.executor.SchemaObject;
-import java.io.IOException;
 
 /*
  * All the commands that needs SortClause will have to implement this.
@@ -18,10 +17,6 @@ public interface Sortable {
 
   default SortClause sortClause(SchemaObject schema) {
     SortSpec spec = sortSpec();
-    try {
-      return (spec == null) ? SortClause.empty() : spec.toSortClause(schema);
-    } catch (IOException e) {
-      throw new IllegalArgumentException(e);
-    }
+    return (spec == null) ? SortClause.empty() : spec.toSortClause(schema);
   }
 }
